Change AssetType to contain a generic type and connector
The asset type is now a struct that identifies the asset by a connector
(where it originates) and an asset kind (help on naming here would be
appreciated) that identifies the kind of asset it is within the
connector. I updated all the right code locations to reflect the change.

Test Plan:

cargo test still functions, cargo run can still run explore and fetch
(sometimes not on nightly, so beware of any failures there for now).