Fix dbt lineage
- Translate the child map from dbt references to relation names.
- Add a `--fetch_first` flag to the `explore` command for convenience
  when launching explore.
- Make sure edges are still added even if nodes are filtered from
  addition to the graph.

Test Plan:

`dbt docs serve` to start dbt docs server.

View
http://127.0.0.1:8080/#!/model/model.dbt_setup.iris_petal_info%23depends_on
to see the lineage for the `IRIS_PETAL_INFO` table.

"Referenced by" is a list of downstream assets.
"Depends on" is a list of upstream assets.

<paste image here>

`jetty explore` to start explore server.

Go to
http://localhost:3000/#/asset/snowflake%3A%2F%2Fcea26391.snowflakecomputing.com%2FJETTY_TEST_DB%2FRAW%2FIRIS_PETAL_INFO/lineage
to see the same table's lineage page in Jetty.

The same assets show up in the downstream and upstream lists, in
addition to some tableau lineage paths.

<paste image here>