Add human panic
This will only happen for release builds, but it gives us a starting
point. We will probably want to tailor the messaging and handling in the
future.

- add human-panic crate
- add `setup_panic!` macro invocation in main

Test Plan:

Added manual panic invocation in main after setup, ran a release build,
and the output is as expected:

```
Well, this is embarrassing.

Jetty CLI had a problem and crashed. To help us diagnose the problem you can send us a crash report.

We have generated a report file at "/var/folders/ty/9ggfl7153_l_8qmllxlg8tgh0000gp/T/report-9e372e8f-d5b8-4f69-9606-9c720a9561d7.toml". Submit an issue or email with the subject of "Jetty CLI Crash Report" and include the report as an attachment.

- Homepage: get-jetty.com- Authors: Jetty Support <support@get-jetty.com>

We take privacy seriously, and do not perform any automated error collection. In order to improve the software, we rely on people to submit reports.

Thank you kindly!
```